move some functions to elcupdater package
